+++ packages/data-table/src/index.ts
@@ -33,3 +33,6 @@ export {
 } from './DataTableArrowPaginated';
 export {QueryDataTableActionsMenu} from './QueryDataTableActionsMenu';
 export {makePagedQuery} from './utils';
+
+// Re-export useful types from @tanstack/react-table
+export type {RowSelectionState} from '@tanstack/react-table';